refactor(docs-retrieval): extract fetch-raw transport core from content-fetcher
Move the network-facing transport core out of content-fetcher.ts into a
focused sibling module `content-fetcher/fetch-raw.ts` (PR 8 of the
content-fetcher modularization stack). content-fetcher.ts drops to ~578 LOC —
`fetchContent` is now a thin orchestrator over the transport + the loaders.

## What moved (verbatim, no logic change)

- `FetchError` + `FetchRawResult` types
- `enforceHttps` — the HTTPS-only security gate
- `generateUserFriendlyError` — FetchError → user-facing message mapping
- `tryUrlVariations` — ordered candidate loop (interactive-learning URLs)
- `tryGrafanaDocsContentLadder` — the content.json→unstyled.html ladder (PR 7)
- `fetchRawHtml` — the primary fetch, redirect trust re-validation, the manual
  3xx redirect branch, and error classification

## Barrel stability

All of this was module-private and only used internally by `fetchContent`,
which now imports `enforceHttps`, `fetchRawHtml`, and `generateUserFriendlyError`
back. No barrel change, no consumer touched. content-fetcher.ts sheds its now
unused imports: `DEFAULT_CONTENT_FETCH_TIMEOUT`, four security helpers
(`isAllowedContentUrl`, `isGrafanaDocsUrl`, `isLocalhostUrl`,
`isInteractiveLearningUrl`), and three url-utils helpers (`isJsonContentUrl`,
`generateInteractiveLearningVariations`, `getContentUrls`) — all now imported
by fetch-raw.ts instead. It keeps `parseUrlSafely` + `isTrustedFinalUrl`
(determineContentType / the entry trust gate) and `generateUrlId`
(wrapContentAsJsonGuide).

## Tests (+19, gap #1 — top priority)

The manual 3xx redirect branch was the most complex, most security-sensitive
untested code in the file. New `fetch-raw.test.ts` pins it directly:
- follows a relative same-origin redirect to a trusted target
- does NOT follow an absolute (non-`/`) Location — surfaces the redirect error,
  target never fetched
- 3xx with no Location header
- redirect target fetch throws / returns empty (error retained)
Plus trust re-validation (ok response redirected to an untrusted final URL),
error classification (404→not-found, 500→server-error, abort→timeout,
Failed-to-fetch→network), `enforceHttps` (https ok / http rejected / malformed),
and the full `generateUserFriendlyError` mapping table.

## Reviewer note

The cross-tier surface is unchanged — fetch-raw.ts is Tier 2 and imports only
downward (types, constants, security, utils) plus its sibling url-utils, so
`architecture.test.ts` stays green with no new allowlist entry.
